requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database (async PostgreSQL)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis (sessions/cache)
redis>=5.0.0

# Config & validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# HTTP client (for testing)
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
